fix(combat): resolve kotlin compilation errors in fight bot and raycaster

+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/combat/killaura/features/KillAuraFightBot.kt
@@ -169,7 +169,7 @@ object KillAuraFightBot : NavigationBaseValueGroup<CombatContext>(ModuleKillAura
         )
 
         val distance = if (context.combatTarget != null) {
-            player.distanceTo(context.combatTarget.entity)
+            player.distanceTo(context.combatTarget.entity).toDouble()
         } else {
             context.playerPosition.distanceTo(destination)
         }
